+++ docs/html/search/all_3.js
@@ -1,4 +1,6 @@
 var searchData=
 [
-  ['fpstr_26',['FPSTR',['../compat_8h.html#aaa60649c7ffe7ed1fbe16dc20ed7e8c3',1,'compat.h']]]
+  ['extern_5fcoroutine_28',['EXTERN_COROUTINE',['../Coroutine_8h.html#a79f1cf897012bbc4df5565f61adc3028',1,'Coroutine.h']]],
+  ['extern_5fcoroutine1_29',['EXTERN_COROUTINE1',['../Coroutine_8h.html#ab92199c3991124cbb6e5e48e27827dae',1,'Coroutine.h']]],
+  ['extern_5fcoroutine2_30',['EXTERN_COROUTINE2',['../Coroutine_8h.html#af52549e71d4a6f7bd18c4688042776b6',1,'Coroutine.h']]]
 ];