+++ docs/html/search/all_4.js
@@ -1,7 +1,4 @@
 var searchData=
 [
-  ['get_5fcoroutine_27',['GET_COROUTINE',['../Coroutine_8h.html#a29e8a48695c9c365bdd589c061afba99',1,'Coroutine.h']]],
-  ['get_5fextern_5fcoroutine_28',['GET_EXTERN_COROUTINE',['../Coroutine_8h.html#ab7a5384c8d2a434781a52a7e9b45c24d',1,'Coroutine.h']]],
-  ['getjump_29',['getJump',['../classace__routine_1_1CoroutineTemplate.html#a84d4aaee5d3ba3750e78574908f2ba11',1,'ace_routine::CoroutineTemplate']]],
-  ['getstatus_30',['getStatus',['../classace__routine_1_1CoroutineTemplate.html#ae849b9399fefd59231c31c96c3ee19cf',1,'ace_routine::CoroutineTemplate']]]
+  ['fpstr_31',['FPSTR',['../compat_8h.html#aaa60649c7ffe7ed1fbe16dc20ed7e8c3',1,'compat.h']]]
 ];