+++ docs/html/search/functions_0.js
@@ -1,8 +1,12 @@
 var searchData=
 [
-  ['channel_85',['Channel',['../classace__routine_1_1Channel.html#ae004356c45d9017d0e375d4468b094e2',1,'ace_routine::Channel']]],
-  ['coroutinemicros_86',['coroutineMicros',['../classace__routine_1_1CoroutineTemplate.html#ad4ac87c4afab0e2e676a3e197dc74c72',1,'ace_routine::CoroutineTemplate']]],
-  ['coroutinemillis_87',['coroutineMillis',['../classace__routine_1_1CoroutineTemplate.html#a658e30e034968b53105f00e1f94d9ed8',1,'ace_routine::CoroutineTemplate']]],
-  ['coroutineseconds_88',['coroutineSeconds',['../classace__routine_1_1CoroutineTemplate.html#a40865b82184d0e0ff93fd9971ef8afed',1,'ace_routine::CoroutineTemplate']]],
-  ['coroutinetemplate_89',['CoroutineTemplate',['../classace__routine_1_1CoroutineTemplate.html#a5bd84b037fb04d15d3842cd59caf3a22',1,'ace_routine::CoroutineTemplate']]]
+  ['channel_119',['Channel',['../classace__routine_1_1Channel.html#ae004356c45d9017d0e375d4468b094e2',1,'ace_routine::Channel']]],
+  ['clear_120',['clear',['../classace__routine_1_1LogBinProfilerTemplate.html#a2f2cf3f7602ff62ffd813650f5711673',1,'ace_routine::LogBinProfilerTemplate']]],
+  ['clearprofilers_121',['clearProfilers',['../classace__routine_1_1LogBinProfilerTemplate.html#a383e87c1dededb9205a5357001d832e9',1,'ace_routine::LogBinProfilerTemplate']]],
+  ['coroutinemicros_122',['coroutineMicros',['../classace__routine_1_1CoroutineTemplate.html#a41acb66b780bc0153117915024811949',1,'ace_routine::CoroutineTemplate']]],
+  ['coroutinemillis_123',['coroutineMillis',['../classace__routine_1_1CoroutineTemplate.html#ad679cad7fc488bc15145d484b6e61364',1,'ace_routine::CoroutineTemplate']]],
+  ['coroutineprofiler_124',['CoroutineProfiler',['../classace__routine_1_1CoroutineProfiler.html#a91a16875699abfda9383b9f29e4f040f',1,'ace_routine::CoroutineProfiler']]],
+  ['coroutineseconds_125',['coroutineSeconds',['../classace__routine_1_1CoroutineTemplate.html#af25c1de1f753912444725f899dfffcb3',1,'ace_routine::CoroutineTemplate']]],
+  ['coroutinetemplate_126',['CoroutineTemplate',['../classace__routine_1_1CoroutineTemplate.html#aebb72dfa544cba3ebcdd5d7647ca8255',1,'ace_routine::CoroutineTemplate']]],
+  ['createprofilers_127',['createProfilers',['../classace__routine_1_1LogBinProfilerTemplate.html#ae8a4228ae33f977c6fe3584eee71beec',1,'ace_routine::LogBinProfilerTemplate']]]
 ];