+++ docs/html/search/functions_1.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['getjump_90',['getJump',['../classace__routine_1_1CoroutineTemplate.html#a84d4aaee5d3ba3750e78574908f2ba11',1,'ace_routine::CoroutineTemplate']]],
-  ['getstatus_91',['getStatus',['../classace__routine_1_1CoroutineTemplate.html#ae849b9399fefd59231c31c96c3ee19cf',1,'ace_routine::CoroutineTemplate']]]
+  ['deleteprofilers_128',['deleteProfilers',['../classace__routine_1_1LogBinProfilerTemplate.html#aa17d5b61420362896fb96e03ed5195aa',1,'ace_routine::LogBinProfilerTemplate']]]
 ];