+++ docs/html/search/functions_3.js
@@ -1,5 +1,13 @@
 var searchData=
 [
-  ['list_102',['list',['../classace__routine_1_1CoroutineSchedulerTemplate.html#adfe184942d2691e1b4c8774623722d8e',1,'ace_routine::CoroutineSchedulerTemplate']]],
-  ['loop_103',['loop',['../classace__routine_1_1CoroutineSchedulerTemplate.html#abf192f6066f9a7cc80219461966cba2f',1,'ace_routine::CoroutineSchedulerTemplate']]]
+  ['isdelayexpired_137',['isDelayExpired',['../classace__routine_1_1CoroutineTemplate.html#adab02ac8807017d168b0ba1c5cec3920',1,'ace_routine::CoroutineTemplate']]],
+  ['isdelaying_138',['isDelaying',['../classace__routine_1_1CoroutineTemplate.html#a731663b0141022c126e9368e1dfe3e17',1,'ace_routine::CoroutineTemplate']]],
+  ['isdelaymicrosexpired_139',['isDelayMicrosExpired',['../classace__routine_1_1CoroutineTemplate.html#acd7d7d1d81f507a04e7a9db8b8820381',1,'ace_routine::CoroutineTemplate']]],
+  ['isdelaysecondsexpired_140',['isDelaySecondsExpired',['../classace__routine_1_1CoroutineTemplate.html#a122e9d35f313975df73ac49afaaa3015',1,'ace_routine::CoroutineTemplate']]],
+  ['isdone_141',['isDone',['../classace__routine_1_1CoroutineTemplate.html#ad753c3b474c659fa6684bde0bd20919c',1,'ace_routine::CoroutineTemplate']]],
+  ['isending_142',['isEnding',['../classace__routine_1_1CoroutineTemplate.html#ad7d690556138014c63be958b0a125516',1,'ace_routine::CoroutineTemplate']]],
+  ['isrunning_143',['isRunning',['../classace__routine_1_1CoroutineTemplate.html#a733385866face30c4fc4223780c1b47e',1,'ace_routine::CoroutineTemplate']]],
+  ['issuspended_144',['isSuspended',['../classace__routine_1_1CoroutineTemplate.html#aa0f33aa5a6197aa143e86708d7f6fb15',1,'ace_routine::CoroutineTemplate']]],
+  ['isterminated_145',['isTerminated',['../classace__routine_1_1CoroutineTemplate.html#a5c74778815c76301895765e7485b4bf8',1,'ace_routine::CoroutineTemplate']]],
+  ['isyielding_146',['isYielding',['../classace__routine_1_1CoroutineTemplate.html#a9c07ae01d589335fa5728210de000755',1,'ace_routine::CoroutineTemplate']]]
 ];