+++ docs/html/search/functions_6.js
@@ -1,19 +1,5 @@
 var searchData=
 [
-  ['seconds_110',['seconds',['../classace__routine_1_1ClockInterface.html#a03c0d39e0167f8a8e7c41f09d19d2861',1,'ace_routine::ClockInterface']]],
-  ['setdelaying_111',['setDelaying',['../classace__routine_1_1CoroutineTemplate.html#a7a8d70181ccc338b1290d42515fd38df',1,'ace_routine::CoroutineTemplate']]],
-  ['setdelaymicros_112',['setDelayMicros',['../classace__routine_1_1CoroutineTemplate.html#ab74b193e9633878d90bd6702323f7f7b',1,'ace_routine::CoroutineTemplate']]],
-  ['setdelaymillis_113',['setDelayMillis',['../classace__routine_1_1CoroutineTemplate.html#a93165b1090754d1d62f8816dfdfe05b4',1,'ace_routine::CoroutineTemplate']]],
-  ['setdelayseconds_114',['setDelaySeconds',['../classace__routine_1_1CoroutineTemplate.html#a12984cedb7d5183a69d21fc49e7d637b',1,'ace_routine::CoroutineTemplate']]],
-  ['setending_115',['setEnding',['../classace__routine_1_1CoroutineTemplate.html#aa62660adbbd4dd79b1534c98b997f11f',1,'ace_routine::CoroutineTemplate']]],
-  ['setjump_116',['setJump',['../classace__routine_1_1CoroutineTemplate.html#ad2f9173ec47abb7bc44356376b84310b',1,'ace_routine::CoroutineTemplate']]],
-  ['setrunning_117',['setRunning',['../classace__routine_1_1CoroutineTemplate.html#a4e4cf5b9c25b340e040578a57382a70c',1,'ace_routine::CoroutineTemplate']]],
-  ['setterminated_118',['setTerminated',['../classace__routine_1_1CoroutineTemplate.html#a76bbbd068a1a8a4e762e7834d28f38ed',1,'ace_routine::CoroutineTemplate']]],
-  ['setup_119',['setup',['../classace__routine_1_1CoroutineSchedulerTemplate.html#ac792080cdec6db9f6539a446d7af587c',1,'ace_routine::CoroutineSchedulerTemplate']]],
-  ['setupcoroutine_120',['setupCoroutine',['../classace__routine_1_1CoroutineTemplate.html#adcc2990d5ab0c53b33ddedc9c6cee579',1,'ace_routine::CoroutineTemplate::setupCoroutine()'],['../classace__routine_1_1CoroutineTemplate.html#ab386f5cb35ee03f927c72c281fe3518b',1,'ace_routine::CoroutineTemplate::setupCoroutine(const char *) ACE_ROUTINE_DEPRECATED'],['../classace__routine_1_1CoroutineTemplate.html#ab24d05e1a1c581a40cce8d4a85695aee',1,'ace_routine::CoroutineTemplate::setupCoroutine(const __FlashStringHelper *) ACE_ROUTINE_DEPRECATED']]],
-  ['setupcoroutines_121',['setupCoroutines',['../classace__routine_1_1CoroutineSchedulerTemplate.html#a86ebd507e675b913301f92af0a34abde',1,'ace_routine::CoroutineSchedulerTemplate']]],
-  ['setvalue_122',['setValue',['../classace__routine_1_1Channel.html#ab950ccdf77fdfaeea67ac19031733229',1,'ace_routine::Channel']]],
-  ['setyielding_123',['setYielding',['../classace__routine_1_1CoroutineTemplate.html#adf41ef1dfedfb5bb2fcd3ff1285bc100',1,'ace_routine::CoroutineTemplate']]],
-  ['statusprintto_124',['statusPrintTo',['../classace__routine_1_1CoroutineTemplate.html#a4429565a460dd3fe53a61ecf60557a76',1,'ace_routine::CoroutineTemplate']]],
-  ['suspend_125',['suspend',['../classace__routine_1_1CoroutineTemplate.html#af106a49c67a4c2920bdc1ac527bad7d5',1,'ace_routine::CoroutineTemplate']]]
+  ['printnameto_153',['printNameTo',['../classace__routine_1_1CoroutineTemplate.html#a62cbb1e87f55ffb06a5ecbc60b3cf361',1,'ace_routine::CoroutineTemplate']]],
+  ['printto_154',['printTo',['../classace__routine_1_1LogBinJsonRendererTemplate.html#aec5865839acf21c5460042b4d2c4aa1e',1,'ace_routine::LogBinJsonRendererTemplate::printTo()'],['../classace__routine_1_1LogBinTableRendererTemplate.html#a1e047b82a91c015425ef4453a863b5c3',1,'ace_routine::LogBinTableRendererTemplate::printTo()']]]
 ];