+++ docs/html/search/searchdata.js
@@ -1,11 +1,11 @@
 var indexSectionsWithContent =
 {
-  0: "acefgiklmrsw~",
-  1: "c",
+  0: "acdefgiklmprsuw~",
+  1: "cl",
   2: "c",
-  3: "cgilmrsw~",
+  3: "cdgilmprsuw~",
   4: "km",
-  5: "cs",
+  5: "cps",
   6: "acefg",
   7: "a"
 };