+++ docs/html/search/typedefs_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['coroutine_139',['Coroutine',['../Coroutine_8h.html#a9971a0ed5c7ba49921d0aa2dbc8f83b0',1,'ace_routine']]]
+  ['coroutine_200',['Coroutine',['../Coroutine_8h.html#a25a42d39c252ea6377c0845010a38c2b',1,'ace_routine']]]
 ];