+++ examples/MemoryBenchmark/validate_using_epoxy_duino.sh
@@ -8,7 +8,7 @@
 set -eu
 
 PROGRAM_NAME='MemoryBenchmark.ino'
-NUM_FEATURES=20  # excluding FEATURE_BASELINE
+NUM_FEATURES=27  # excluding FEATURE_BASELINE
 temp_out_file=
 
 function cleanup() {